backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Scraping
playwright==1.49.0
playwright-stealth==1.0.6

# Anti-detection
fake-useragent==2.0.3

# Export
openpyxl==3.1.5

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
httpx==0.28.1